[#172989345] refactoring
